ViewTransformer to compose
