+++ mvlearn/__init__.py
@@ -4,7 +4,6 @@ import mvlearn.semi_supervised
 import mvlearn.datasets
 import mvlearn.utils
 import mvlearn.decomposition
-import mvlearn.preprocessing
 import mvlearn.compose
 import mvlearn.model_selection
 import mvlearn.plotting  # noqa
