--- mvlearn/preprocessing/__init__.py
@@ -1,5 +1,0 @@
-from .repeat import ViewTransformer
-
-__all__ = [
-    "ViewTransformer",
-]
